+++ resources/views/layouts/navigation.blade.php
@@ -26,7 +26,7 @@ $showBirthdayListConfig = \App\Models\Configuration::getBool(\App\Models\Configu
                     <a href="{{ route('dashboard') }}">{{-- add setting to show or hide app icon --}}
                         <x-application-logo class="block h-9 w-auto fill-current text-gray-800"/>
                     </a>
-                    @php($appName = \App\Models\Configuration::getString(\App\Models\ConfigurationKey::APPEARANCE_APP_NAME, default: config('app.name', 'Club management'))) {{-- add setting to show or hide app name --}}
+                    @php($appName = \App\Models\Configuration::getString(\App\Models\ConfigurationKey::APPEARANCE_APP_NAME, default: config('app.name'))) {{-- add setting to show or hide app name --}}
                     @if($appName)
                         <div class="grow text-center text-gray-700">
                             {{ $appName }}