+++ CHANGELOG.md
@@ -3,9 +3,12 @@
 All notable changes to this project will be documented in this file.
 
 ## Unpublished
+### Added
+- appearance setting part (set application name, and logo)
 ### Changed
 - attendance display group list switcher is frontend managed again, removed member filter
 - add today switch for event filter, update accordion component
+- change default logo from a club specific to font awesome based logo
 ### Fixed
 - attendance display and record render works now after group/list switch